Developed the Article Page
The ArticlePage pulls data from the Article file which pulls from the TempDataSource as well. When you click on an article card it will take you to the full article.

I also had to change my function that changes the article card color when clicked so now it remains the changed color when you go back to the feed.