Edit Account Screen and Support Screen
Account screen is editable. I did not write the mutable state to be saved as I believe we want this to work in conjunction with supabase. For now, the aesthetics and structure are in place.

I also wrote in a Support screen. There is no real Customer support so I threw in a graphic and some placeholder text.