Adding logic to sign up new users
Also create the user metadata object with filters to store on the user and added methods that should work to update the user and the user filters when we get to that point.

I am able to sign up with a new account and added validations in the create user view model.

Localized all strings.